Add option for not firing action events on key navigation.
By default a JComboBox will fire an action event everytime the
selectedIndex is changed on the combobox model. Add an option for
CommonComboBox to change the selectedIndex of the list component
instead of the combobox.

This mimics the UIManager option: "ComboBox.noActionOnKeyNavigation"
which is global. We want to support both approaches in AS without
changing the global setting.
This is in preparation for fixing b/162759463

Bug: 162759463
Test: Unit test added.
Change-Id: I9f13d7475a0290d1b9d3b72368ae06b6703f9bb2

GitOrigin-RevId: 2f20a7fcfa41307c355f78a006bfab8ee931ceea