Remove hack for detecting an update in the popup of a ComboBox
We want to be able to navigate in a ComboBox popup without updating
the underlying value. Some updates can be time consuming and make the
popup navigation slow.

Before this change: a PopupMenuListener was responsible for detecting
when a property value should be updated when the popup is about to
become invisible. This was a hack that is hard to maintain.

With this change: take advantage of the new flag in CommonComboBox:
"actionOnKeyNavigation" which updates the selection in the list instead
of the selection in the ComboBox itself.
Changes are then detected by adding an action listener to the ComboBox,
which is how the control was designed.

Bug: 162759463
Test: Unit tests added (some tests are moved from the model test)
Change-Id: I188677e4b63211d2293157c5889c1bc35b2662e2

GitOrigin-RevId: c99aca895abdf44a9bf5892838599c07582ad0a8