Add placeholder run configuration to mutable list
AndroidProcessChooserDialog#setupDebuggerSelection adds a null element
to a list of current run configurations using List#add. This fails when
there are no run configurations in the current project because the
list type will be immutable, causing a UnsupportedOperationException.

This CL uses an ArrayList to hold the run configurations to guarantee
mutability.

Fixes: 162956657
Test: Clicking on the "Attach Debugger to Android Process" button with
no run configurations present in the project doesn't cause the exception
anymore.

Change-Id: I4801f8c9455903598fc62fb3b4ba6ea443da0445

GitOrigin-RevId: 34eae17567d873235a6eda2d1ffd7203d188a0b8