[AUA] add target(s) for main refactoring
This might seem like a primarily cosmetic change -- the existence of a
"target" for this refactoring only adds a header and a line in the
UsageView preview.  However, it appears that something somewhere in
the UsageView tree builders assumes that there will be a non-zero
number of targets, even though ostensibly zero targets are supported:
there are consistency errors when re-running the refactoring from the
usage view window (the rerun button causes all current usages to
vanish, and new usages found by the processor are not added to the
tree, so rerunning ends up with a blank view; meanwhile, using our
back button and re-previewing adds new usage groups without clearing
out the old ones).

Pending time to investigate the odd behaviour of the UsageView tree,
add as a target the AGP dependency element itself, whether that's a
version declaration or the whole dependency.  Override the
WrappedPsiElement getPresentation method in order to make the
displayed text of the target be something appropriate (the default
methods end up giving the null string for our Psi expressions).

The downside of adding a non-zero number of targets is that their
existence causes a non-functional settings button to appear in the
toolbar ribbon instead.  It is non-destructive (it simply does
nothing) so this is probably an acceptable tradeoff for the moment.

Bug: 163020008
Test: manual
Change-Id: I5a0a1c1dad3ca12b3f4f051318b88daa239291c5

GitOrigin-RevId: 37b535a275723d053e0be777eeb50c162465f0fd