Save and restore tabs each time a database is remove and added
Instead of saving open tabs only when the app inspection session
is over, this change saves tabs each time a database is closed.
Tabs are still restored each time a database is opened.

Tabs associated with no database are restored as soon as any
database is available.

This will make it possible to restore tabs once an offline
database becomes available.

Test: DatabaseInspectorControllerTest.kt
Bug: 162403339
Change-Id: Ie43696648dc569e4f10e5ef0133f69aeec316964

GitOrigin-RevId: 12ba5356338f206ea9bf577e7786aeddf55dfb64