Update retention snapshot file names
Because of an update of the Android Test Retention workflow, we need to
update our naming convention for Retention snapshot files.

For each exported snapshot, it has an internal snapshot name. Previously
we always use test_failure as its internal snapshot name. But because of
the update of our workflow, we start to have different internal snapshot
names.

To tell Studio about the internal name, we encode it into the exported
snapshot file name. The internal name is always "failure[0-9]+". This
commit parses the internal name from the exported snapshot name.

Bug: 162996605
Test: UtpTestResultAdapterTest
Change-Id: If4bf1ad4feea982e95d89e7876ca966355201040

GitOrigin-RevId: bdda76ef7b3d2e868139d701fe975905dbf09dd8