[AUA] entry point from lint deprecated configuration check
One of the optional project upgrades to go from earlier versions to
4.2 is/will be the replacement of deprecated compile/runtime
dependency configurations with
implementation/api/runtimeOnly ("optional" in the sense that they are
deprecated but not yet removed: scheduled for removal in 5.0).

The optionality means that it's awkward to show it to users in the
main upgrade flow (probably they just want to get on with things) but
the imminent removal means that strongly encouraging users to do the
upgrade is wise.  So, implement another entry point to do project-wide
refactoring of deprecated configurations using the AUA machinery from
the lint quickfixes for the deprecated configurations.

Bug: 110017227
Bug: 155123172
Test: dialog behaviour tested
Test: manual test for invoking the dialog
Change-Id: I618bfd1d9afa7e5afc2d5a658eda042bf41e896e