Add support to build the studio updater from prebuilts
The prebuilts now include the vanilla IJ updater.jar,
that we use as a base for our updater. The updater could
be organized better (by not having to specify all the runtime
dependencies manually), but I'm mirroring what we currently
do in tools/idea. We can further improve this once we
switch.

Test: There aren't. Filed b/163193375
Fixes: 163172102
Change-Id: Ib9f54d1e01fd6345f440ad4d7434576c7e7efd5d

GitOrigin-RevId: d9fe39f457c6cb9bfff4804b8af33e2e466eb0c2