Fix constraint_referenced_id="null"
For many constraint helpers, when they are newly created it checks for
selected ids and add them to referenced_id. However, when id does not
exist it adds "null" to the list.

It doesn't compile, and causes users to do extra work of removing "null"
manually.

This CL checks for null so it doesn't get added to the attribute.

Bug: 156771573
Test: Tested with flow, group, barrier and guideline
Change-Id: I29d5b53e6e955a2c78e897b4d5137174da9caf92