IDEA-247946: remove dependency `intellij.android.buildCommon` > `intellij.android.common`
`android.buildCommon` is used in the JPS build process. `android.common` has a lot of dependencies which are not available in JPS (available in the IDE only). In order to avoid potential NoClassDefFoundError, just replace `BufferingFileWrapper` with `FileWrapper` and remove the dependency. There is no performance penalty for this change, because `ProjectProperties.parsePropertyFile` will wrap the `InputStream` with the `BufferedReader`, so we just removed double-buffering (which is good in 3D games, but often useless when reading files)

(cherry picked from commit 906828631bd366e9bd94bc8cff4cedb7a2c98875)

Change-Id: I906828631bd366e9bd94bc8cff4cedb7a2c98875

GitOrigin-RevId: 92744d89c61744022e0b1851236e94b3e432569b