Hide animation inspector icon when there are no animations
We should only display the animation inspector icon in Compose Preview
toolbars when the Composable being previewed has animations to be inspected,
otherwise it might just confuse users.

This CL introduces the SceneUpdateListener to SceneManager, adding the
possibility to customize the scene and performing additional operations
based on its root component when updating it. For Compose, we update the
animation inspector visibility in the preview toolbar as mentioned above.

Bug: 163469060
Test: Manual by previewing Composables with and without animations.
Change-Id: Id090b8c7f391492453894b7ef2a274dea9159eaa