Redo the implementation of TreeImpl
Previously the badges were forced to the right of the component
tree by overriding the preferred size of each row in the tree.
That approach caused problems when a horizontal scrollbar was
added: the tree could grow but not shrink in width.

With this new approach the preferred widths are computed as it
normally would. But the badges are placed at the right edge of
the tree.

Also the hittest / mouse click / tree expansion handler is
updated to work with this approach.

Fixes: 162767339
Test: Existing tests should pass
Change-Id: I3f5048046ac42e38c56809a38c6da4b0c3f4184b

GitOrigin-RevId: 5cb3f97d5d757703283b3f637069465ff9019fd1