Enable SantaTrackerKotlin highlighting benchmark
This reverts commit 0b293edc94135172d1c62ab97eb0f70327e89350.

The benchmark may still be flaky, but Kotlin 1.4 is coming soon and
it will be valuable to have benchmark data during the transition.
Furthermore, it will be useful to know whether Kotlin 1.4 reduces
flakiness.

If the benchmark is still flaky after Kotlin 1.4, we can consider
other mitigations such as reducing the scope of the benchmark to only
the 'core' syntax highlighting pass rather than all inspections, or
working with JetBrains to investigate the underlying issues.

Bug: 149849795
Test: passes locally
Change-Id: Ie0fabe63034aa390ac14f7e10a43945de14518f4