Change HTreeChartTest to use metrics instead of magic number
The font metrics change by 1pt between linux/mac and windows.
The hardcoded values in the test assumed the font size and
padding to compute total height. This change computes the
height based on the formula used inside HTreeChart.
The pro here is we have accurate accounting across all
platforms and future jdk changes. The con is if the font
size changes significantly only a manual verification will
catch it.

Test: fixed
Fixes: 163140294
Change-Id: I2a44ebbf343b1b71611a4a6ed14a3a02cc830c84

GitOrigin-RevId: 3683b80c2cc60650faac800c18cfb2fb6213ea22