Prepare source providers for being lazy
The existing implementation of `[Named]IdeaSourceProviders` stores
all its entries in a form of file urls and includes all pre-configured
location which not necessarily exists at the time of configuration. This
data structure becomes huge for larger Android projects with multiple
build variants.

This change is a step is to enable implementations retrieving lists
of urls/files on demand from the underlying storage without the need
to keep file urls in the memory all the time.

Bug: n/a
Test: n/a
Change-Id: Idaf49f522fea36d3e1260a0c044bb7c2c4b7d15e