Fix problem with multiple processes in process dropdown
We were relying on process start/stop events to determing the
actual running processes on a device. However the process stop
events were often lost making it look like the process was still
running.

The new approach is to use process start/stop events to invalidate
the currently cached set of processes and then ask for all live
processes as a delayed response to an invalidation.

Fixes: 158459958
Test: Existing tests should pass
Change-Id: Ie5f6315b08dee9f923e7cba9a2d1c1fa802a4d25