Make IdeaSourceProviders implementation delegating to their data source
The delegating implementation reduces the memory usage of large project
with multiple build variants significantly. If it turns out that the
processing overhead is too high for resolved property variants we
will memorize their values (except the manifest file which is not a
project root) as source providers are discarded on each project roots
changed events.

Bug: n/a
Test: source provider snapshots
Change-Id: I4fcdbc1dc71146788bbdc4088da05c45a70ea02d

GitOrigin-RevId: 5a70c56e76e19bfadcdf65925c66ddc923561950