Add intellij_platform rule
This change adds the intellij_platform rule that
can handle embedded plugins. These plugins are processed
and their plugin.xml stamped with the build id.

Note: When adding this rule it failed to process
the "Compose" directory inside the sdk, as it does
not contain a plugin.xml. This will be fixed separately.

Test: N/A
Bug: 156497102
Change-Id: Ie28d8ba5e5f3c5d2b97a3339e6e6fcfc51329023

GitOrigin-RevId: 1a14d91dbae75381951faadae334e7aa53f13201