Fix do not restore tabs for in-memory dbs
In-memory dbs have path and name with this formant
":memory: { somehascode }", not just ":memory:"
therefore the previous filtering was broken.

This CL also adds an extension function to SqliteDatabaseId
to know if the db is in memory or not, so we don't have
to remember the format of the path.

Test: intellij.android.debuggers.tests
Bug: 163506564
Change-Id: I142657787e7bb9e8f35f88f6b5863fb1fc029d77