Structure AndroidGradleProjectResolver.createAndAttachModelsToDataNode
in a way that (1) we get models from `resolverCtx` which is basically
a lookup in a map (2) prepare models to attach to the data nodes tree
and (3) we attach models to the tree and perform the remaining setup
operations which do not fit into this pattern.

This is to make it possible to prepare some IDE models on the Gradle
side.

Bug: n/a
Test: n/a (existing snapshots)
Change-Id: I57f8099ee76dad3133ad0ca2e2aad506b9b02f93

GitOrigin-RevId: 53b7f6fd110b557b964c8a82af78ac496930fdf8