Transitions between static and interactive previews
Interactive is not a complement of static mode, there are also
two transition modes. Because of that we can enable/disable
interactive mode with InteractivePreviewElementInstance but can
not use it as an indicator. To indicate current preview state
one should use status() method of ComposePreviewManager
interface.

In addition to that, UI should respect transition states as well
and display some elements only in static mode and some only in
interactive (therefore both of these sets should be hidden during
transitions).

Bug: N/A
Test: manually tested
Change-Id: I24016b615d512258d060abb588748dbdc3271f94