Use adb serial number to identify devices
To download files using DeviceFileDownloaderService
we need to identify the target device first.
The reliable way to do this is to use adb serial
number.

This change exposes the device serial number from DeviceFileSystem
and uses it in DeviceFileDownloader service to find the
correct device.

Just using the name as we did previously does not work,
becase the name is just a human readable string, can't
be reliably used as an identifier.

Test: NA
Bug: 160868563
Change-Id: I53228ee24c2326feefc4c13c524d1f01154e79bf

GitOrigin-RevId: e1bf6245c718cbfdd081f80b41cce32d401a1b75