[Dsl] parse plugins with versions
In neither case (Kotlin or Groovy) is this a full solution: the plugin
model depends on having particular shapes of Dsl elements, and
arranging for the parser to produce those shapes (or modifying the
model) is not completely straightforward.  Unfortunately, plugin
declarations with versions are being seen in the wild, and some of our
other functionality (e.g. the Project Structure Dialog, the Upgrade
Assistant) depend on being able to identify plugins used in a Gradle
project.

For now, parse slightly lossily: identify the plugin, and throw away
the version (and the Dsl representation's association with the version
declaration, unfortunately).  This is good enough to handle our
current uses.

Bug: 165576187
Test: added parsing tests
Change-Id: Ie44305e5759802362ce225015d23d8e904b57817

GitOrigin-RevId: 8779c2acd14fc284a0dac629dce973020b640c5e