[Dsl] handle project injections in the model
If a property value references a project (e.g. using `project` or
`rootProject`), that property will have a dependency on a project
property.

In general, properties have holders, considered as the context for
resolving them.  In almost all cases, this is the Dsl element's
parent (which is itself a known problem, conflating lexical scope with
structural object representation); however, a project is represented
by a GradleDslFile, which has no parent.

Smooth over this mismatch by defining the holder of a project property
to be its own Dsl element (which will be the right context for
resolution in any case).

Fixes: 165017523
Test: existing test strengthened
Change-Id: Ic357e4b64dbff91d9c12781bff7e968a25df47d3