Separate requesting and delivery of Android models
This is a prerequisite to moving conversion to IDE models to the Gradle
process. This change does not change which models and in which order are
requested and should not have any visible effect.

After this change we will collect all Android models in `AndroidModule`
instances before passing them to `AllModels` consumer which later
serializes all models fetched from the Gradle process and delivers them
to the IDE process.

Note that with this change we will not request project sync issues
from non-Android projects anymore as they should never be available for
such projects.

Bug: n/a
Test: existing sync tests and synced project snapshots
Change-Id: Ie7bc177f78e8cf357ef9c8d4c597c0dd3e3cbdba