Fixes extraction of android-studio in test
The previous extraction was losing all the executable attributes,
due to https://bugs.python.org/issue15795,
and I was setting it back on studio.sh. However, jre/bin/java
remained non-executable, and made the script not find the
embeded JDK, falling back to the system one.

Which it actually found (py_test sandbox issue?), and triggered
a different issue that caused it not to find a class.

This change fixes that.

Test: this is the test
Fixes: 165138879
Change-Id: Icb9e797373ffcf122451ab0669d7bcba3d8272be

GitOrigin-RevId: caa0f13c4fc004e33189400d9e96dabd0da33c3f