[Dsl] buildSrc should be part of the build model
We cannot model everything under buildSrc (after all, it's even more
arbitrary code than the arbitrary code present in Gradle build files),
but we should be able to model the Gradle build files used to control
the compilation of buildSrc itself, if any.

This is in fact necessary in order to properly upgrade project build
files, for example if the build files in buildSrc themselves depend on
the Android Gradle Plugin.

Composite models include the buildSrc of the including project, and
not that for any buildSrc which happens to be lurking in their own
root.

Bug: 163364829
Test: existing tests adjusted, added ProjectBuildModel test
Change-Id: I62928000422f36e9f9b5b45398ca511a5ed49811

GitOrigin-RevId: 2e9732232e10b1e2bfbed62a11df4ec7f54799e8