Do not change Java home after calling prepare
... when running sync, to have consistent ways of finding what JDK to
use for Gradle.

When running sync, we configure the project settings using
GradleExecutionHelper#prepare but then we change the java home setting,
causing discrepancies between sync and build (also, we potentially log a
different path to what is actually used). This method already receives a
GradleExecutionSettings with the java home, and it contains already the
expected value.

This CL removes the extra call to operation.setJavaHome(javaHome)

Bug: 166774744
Test: Existing tests that change JDK, but we need to add some integration
tests that cover this. b/167414916 traks those tests.

Change-Id: I61063712f4bdac0df1cd6114423272dccb63480a

GitOrigin-RevId: 6bde4ffa6c8da473b17ab2d33712b70278be3fc3