Add view binding to the BuildFeaturesModel for AGP 4.0+
Also add setViewBinding to the RecipeExecutor for AGP 3.6.

buildFeatures block needed to be added when an Activity is added instead
of a module is newly created because an Activity may be added to an
existing project without view binding support.

Thus moved adding the build features logic to RecipeExecutor and call it
from Activity's recipe

Bug: 161989148
Test: TemplateTest, UI tests, manually created Basic Activity with AGP 4.2 and 3.6
Change-Id: I8b1ab6f07af51a2c1d3a3c2e2fc17748c790b0c6