[AUA] detect about-to-fail upgrades
If we fail to find any usages from the classpath refactoring
processor, then our attempts to upgrade (whether recommended or
forced) will fail.  In the particular case of a forced upgrade, that's
quite annoying, because the subsequent sync will pop up another modal
dialog, but it's also annoying in the recommended case where the
balloon notification will come back time and time again.

Detect the case where the classpath refactoring is a no-op, and exit
the upgrade flow with a dialog asking the user to perform necessary
upgrade steps by hand (and giving a clue about what features we
currently know we don't support, specifically code in buildSrc).

Fixes: 167615343
Test: manual
Change-Id: Ica4d8e5ed472d8ce2ab7691ac8fa447045597464

GitOrigin-RevId: 8b4c58b0e658717d349b0380e16e9a01a0df477c