Bump TPD gRPC max inbound msg size
This value is just a sanity check to prevent gRPC from trying to parse a
message that is too big for it. Since we have been adding more and more
data to TPD, the 128 Mb limit is getting more and more strict and until
we move the code to streaming gRPC mode, we need to allow a bigger size.

The potential user is causing OOMs in the JVM, but for traces bigger
than 100 Mb we already display a warning when loading the trace.

Bug: 160917201
Test: n/a
Change-Id: I7ae6958feda1f22dfeab29c633186710254eab8b

GitOrigin-RevId: 36d4f7c27262af5ff93c47f820a6754fa8c8c654