[AUA] specific field for back-from-preview action
We might have arbitrary actions added to the preview button line, but
some of them might have special meaning.  The one that we currently do
conditionally use (in the Java8 dialog, but not in the CompileRuntime
entry point) is special in the sense that it is a good approximation
to a ShowSettings() action, so being able to inquire about its
presence or absence directly, and get hold of it if present, will be
useful.

No other change in functionality; use the setter of the new field to
maintain the additionalPreviewActions list.

Bug: 168057691
Test: existing
Change-Id: I6d36cd4a67bd948c9fd2839cd15b0648d353e78f

GitOrigin-RevId: e09bda303971188e6bb7d6e7e9ace1da4f72db97