[AUA] better behaviour of settings button
The preview shows a settings button when the target of the
refactoring's usageViewDescriptor is a ConfigurableUsageTarget, which
the results of PsiElement2TargetAdapter.convert() always are.  That
settings button does nothing (in the case of an AGP upgrade) or might
open a findUsages dialog (in other upgrade cases) -- neither of which
is the right thing to do.

Wrap the PsiElement2UsageTargetAdapter objects in our
own (mostly-delegating) implementations of the UsageTarget classes,
Configurable- or not as appropriate to the particular Upgrade
Assistant invokation: if we have a back-to-dialog action, use that,
otherwise make the usage target non-configurable.

Bug: 168057691
Test: manual
Change-Id: Ie681ef1f1b59827865838e4f77e694753fee7146