Add jdk 8 to ProjectJdkTable for tests
... that use jdk 8.

Tests are run using JDK 11 and the instance of ProjectJdkTable only
contains jdk 11. Since the test project uses jdk 8, the tests cannot
find a valid JDK in the ProjectJdkTable and this causes a
jdkConfigurationException.

This change adds jdk 8 to the ProjectJdkTable before loading the project
so the expected jdk is present.

Bug: 166639453
Test: OpenProjectIntegrationTest
Change-Id: I63e0ba120108bfc93ae66e05976ebf0f0aebda29

GitOrigin-RevId: 215bbc807de1c66c8c87dd7e4cdd5bf2dec1322b