Add checkbox for disabling deploy/run optimization
The checkbox will cause Run to always use a package manager
installation, even if the device is compatible with optimistic run and
optimistic run flags are enabled.

Note that even when the checkbox is selected, the deploy cache will be
maintained, in order to support toggling the checkbox seamlessly.

The checkbox will *not* impact Apply Changes or Apply Code Changes; it
should, but the refactoring neccessary to make that work properly
requires its own CL.

Fixes: 168136331
Test: Manual
Change-Id: I073971000c77c53f22faae56c00d2bcf9eb3d3fa

GitOrigin-RevId: ba437d218432bdba02c96d37159a866f8b55c942