Use public test framework classes
Intellij test code is nor part of their test API,
which makes it very unstable code to depend on.
This change uses code that is part of their test
framework APIs.

This will be easy to enforce as we move to prebuilts,
as these test clases are not available and this would
not compile.

Test: this is a test
Bug: N/A
Change-Id: I9918384807a1333e93cd4a1879c995a976a61dc2

GitOrigin-RevId: 01774a48070bff2c96606543d6bff01d8669609c