Use win64 folder for JDK 8
When we started using JDK 11 we stoped having 32 bit version of the JDK
and thus the win32 folder was removed and the win64 folder was renamed
to win. But we still have both folders in JDK 8 from prebuilds.

This change uses the win64 folder instead of win and adds some log
messages to help identifying issues when a jdk location is not valid.

Bug: 165937355
Test: Existing Windows tests that depend on jdk 8.
Change-Id: I7c22375314ac9cc7db392c3d20a1c2400f7e74c2