Sync additional native variants only
When deploying an app with native modules to a device whose ABI is
different from the currently synced in the IDE ABI variant we may need
to fetch additional models from Gradle. The existing implementation
requests full-variant sync to achieve this, but this is both inefficient
and may cause the selection of build variants to be reset.

This change implements a new sync mode which fetches missing native
models only. Moreover, we (1) move the check for missing models to after
build completion. This is required to allow native v2 sync to generate
models while building the app and avoid requesting additional sync. And,
(2) we look for and fetch native models for the required ABIs only.

Bug: 151654753
Test: MakeBeforeRunTaskProviderIntegrationTest enabled
Test: MakeBeforeRunTaskProviderIntegration35Test enabled
Change-Id: Ib8f69235c5528f4b0f260dfd79dff6baca24deb8

GitOrigin-RevId: 1c186d2c9ab613dbe116ac6289cddb73d7583964