Specify if error message is about NDK or SDK
... for validatedSdkPath.

In the PSD we use the result of validatedSdkPath to know if the SDK and
NDK paths are valid but the error messages created by this method not
always have the sdkName parameter on it, making it difficult to know if
the error comes from the NDK or the SDK paths. This change adds the
sdkName parameter in the error message of those cases.

Bug: 168363601
Test: SdkPathsTest modified to check for this change.
Change-Id: I20c0fbe8707367d066d928240bd423d7eb02ffc1

GitOrigin-RevId: 8d435c57c3e071b0cd2d97f97398f2c6e21cd7c4