Open file databases after all downloads are completed
Instead of opening dbs as soon as the file download is done,
we now wait for all files (of all dbs) to be downloaded.

The problem with opening dbs as soon as their file is downloaded
is that it would be possible for the user to open a tab
before all the downloads are completed. This would cause the
download progress indicator to be removed from the UI,
as the mocks currently are designed.

In the future we should reconsider this, move the download
progress indicator somewhere else and open dbs as soon as
the file is ready.

Test: DatabaseInspectorProjectServiceTest.kt
Bug: 168975660
Change-Id: I0c2d829be5686532901fc756b1d15e29f66520b4