Automated deletion of of incompatible snapshots
When incompatible snapshots are detected during opening of the "Manage Snapshots" dialog, the incompatible snapshots are either deleted automatically, deleted after a confirmation, or preserved. This behavior is controlled by an application-level preference. The confirmation dialog contains the "Do this from now on without asking" checkbox, which in combination with the clicked button changes the same preference.

Test: ManageSnapshotsDialogTest
Bug: N/A
Change-Id: I7fd5f7cddd32cf736a2da96eabdf6ce2164f9959

GitOrigin-RevId: 60ccf2ce1bd25b670005bcd2139e11d5bd94fee3