[Dsl,.kts] somewhat clarify convertToExternalTextValue
Rather than appending dereferencing punctuation to a list/map
expression, and then trying to delete that punctuation after the fact
if it turns out that we are not dereferencing it, maintain the
invariant that the `externalName` being accumulated is always
syntactically valid at the end of each iteration through
resolutionElements.  The function still needs more work to be truly
clear, but at least in this one respect it is easier to reason about.

Bug: 155168920
Test: existing
Change-Id: Ibba7b44bfbe8bec98d90fc4dab5d6a103d52869c

GitOrigin-RevId: b988fc7ba915006f02c88aeaca8c18827bd0f7ac