Automated deletion of incompatible snapshots (second attempt)
When incompatible snapshots are detected during opening of the "Manage Snapshots" dialog, the incompatible snapshots are either deleted automatically, deleted after a confirmation, or preserved. This behavior is controlled by an application-level preference. The confirmation dialog contains the "Do this from now on without asking" checkbox, which in combination with the clicked button changes the same preference.

Test: ManageSnapshotsDialogTest
Bug: N/A
Change-Id: I7a197839b63799d154605dc256f662d3ad2bbd09