[AUA] do (much) less parsing on the EDT
Arrange for the parsing steps of the Upgrade Assistant to be done on a
pooled thread.  In the case where we are about to show the user a
dialog, explicitly make the models computed before changing the
modality state.

Rearrange things to make sure that the entry to the various
performUpgrade functions is always on a background thread.

Bug: 169366601
Test: manual
Change-Id: I6f338addb9a32685db510d6ba660f7d10e5a7f4a