Delay agent attach for startup profiling
This change delays when the perfa agent attaches.
If the user is doing a startup cpu / memory profile
the agent is attached after the recording stops.
Otherwise the agent is attached when the app
launches.

JVMTI works by lazy deoptimizing the bytecode of
a class. It does this by signaling only when a
class file is read. This signaling causes an error
inside heapprofd and it stops recording allocations.
By delaying JVMTI actions till after native recording
stops we avoid this issue at the trade off of not
having network / event information while recording.

Test: added
Bug: 160827189
Change-Id: Ie3806daea07711587ebd7f8c7001ba07b794b70c

GitOrigin-RevId: e8e4ebd9243dd57e717bba3c50c9d7a889ebdf50