Changes stamping mechanism
Stamping was done at a file level and that cuased several
tools to be more complicated than they needed to.
Having per file granularity meant that bazel needed to
be aware of all these files at analysis time, and if we
decide to stamp one file more or less, the rules needed
changing.

With this change stamping happens within zip files that
will be used to override the main ones. This unifies how
platform and non-platform plugins are stamped, and removes
the need for a few other tools.

Test: N/A
Bug: N/A
Change-Id: If9ab014edc632ecd6b77d99c560e35d29183c35a

GitOrigin-RevId: 465453d56684b9ee0ff1409e3af22981b9c8d1a5