Check sync perf tests can be synced and built
... in pre submit.

The tests runs sync and build, failing if either of them fail. The test
projects used are the same as those used in perfgate for sync. Due to
the time it takes to build only Average and SantaTracker are build
and only Average is run in Windows for presubmit.

This change also adds Santa Tracker for perfgate.

Bug: 169436495
Test: New tests to be run in presubmit
Change-Id: Ie8bab9e40b9d6781f23db0799fbd262862913dd5

GitOrigin-RevId: 503a9261a6ea832e9792f55bc5e361a83ec19add