Fix race condition on PerfettoParser
The UI allows to trigger to parsing requests very close (< 200
miliseconds), which causes a race condition as both threads tries to
load a new trace file and query it.

So we introduce a cross-instance lock (because CpuCaptureParser builds
new PerfettoParser for each parsing attempt) in order to control this
behaviour.

Test: manual
Bug: 169637961
Change-Id: I7abd88eb623df88deac16135c8e72c9da4b709b3