[BA] Fix Vertical chart scroll refresh
Problem:
In some cases optimization is applied to a scroll pane when on scroll event
instead of rendering everything it just takes the current image, shifts it and
renders only the new part. This optimization can not be applied to a chart part
as it does not move vertically, it has to be re-rendered for each new position.

To solve this the best way is to split tree part and chart part into two
separate viewports (that would still be controlled by a single scrollbar)
and set the chart part '.scrollMode = JViewport.SIMPLE_SCROLL_MODE'.
This also helps solving too long tree content problem, we want only
the tree part scroll horizontally in this case.
It is tricky to do this properly though.
After experiments with different options around it I ended up with the
following solution: add chart part as a row header. Row header is wrapped
to a separate viewport, we can set simple render mode to it problem solved.
It can not be it though, because by default row header is on the left
from the main viewport and we need chart on the right. FLIP can make the
trick, but then scrollbar will be on the left which is also unexpected.
The solution is to tweak the internal layout manager and swap the
tree and chart viewports horizontally. This simple transformation
should not break anything in this specific case.

Bug: 169220062
Test: existing
Change-Id: I952926346cea5173f4c9b260a0d22b9e2cb3eb7d

GitOrigin-RevId: f21fe34b13b65bab09394bef030d7223a0dcc1ed