Add test run count to AndroidTestCase ID in TestMatrix.
By including the run count to the ID, it won't de-duplicate
test runs with the same class name and method name. If
a same test method runs twice, there will be two result
rows in the table. (And this is the same behavior to
the previous Test panel).

Test: DdmlibTestRunListenerAdapterTest
Fixes: 169947443
Change-Id: Idfdec20ab386bc41e0a19d6cd20bd0440076d2bd
(cherry picked from commit 91b11658a1821c5710c93acbc7181507d71c2031)