Count ProcessCanceledException as build cancelled
Currently, when the user cancels the build via the task bar,
IJ throws a ProcessCanceledException (this is normal). However,
GradleTasksExecutorImpl's catch of this exception doesn't count
it as a cancellation, and tries to perform followup processing
that fails when such an exception is thrown. And when the follow
up processing happens, it throws an error that causes a failure
to release a lock that Gradle is holding on to, and thereby
permenantly disables the Run button.

The fix is to count ProcessCanceledException as a build being
cancelled, thus avoiding the problematic followup processing,
and allowing the unlock to proceed.

Fixes: 167977589
Test: manual cancellation of build and check status of Run button
Change-Id: I54ea09714a9a3ae5acfc5b5a5ad49805028f0b52

GitOrigin-RevId: 9642ff58ebb7af1832c41b8a8dd5211b3da4bb05