Fix component splitter in AndroidTestMatrix
JBSplitter and OnePixelSplitter has a property of
minProportion and maxProportion however this value
is ignored when there is no sufficient width such
as 1 px width. The internal proportion value is
corrected after you call setProportion() with
a value that is different from getProportion().

This patch changes to manually check min and
max value and update it properly in doLayout()
method.

Test: tested manually
Fixes: 170234515
Change-Id: I6f6c7800e915f6e5fcf068b54d21f11160d43c6b