Add message panel to new query view
In the new query view we sometimes display a message
to the user. So far we've used the empty text of the
table. This CL improves this solution so that we can
show a panel specifc for showing messages, instead of
using the empty text of the table.

Test: SqliteEvaluatorControllerTest.kt, SqliteEvaluatorViewImplTest.kt
Bug: 169387231
Change-Id: Ibe7fee2f44936b2e84901c15b461be238dd9f503
(cherry picked from commit 00f9a6f2a8d2a0a06a3c5b6cb43074348c28d3e4)