[AUA] update plugin versions in lockstep with Gradle
An AGP upgrade can trigger a Gradle version change, which itself can
render the build incompatible with other Gradle plugins (and in
principle other things, but let's keep it simple-ish for now).

Keep a registry of plugin group/name to function of one argument, a
GradleCompatibleVersion, which returns the lowest compatible version
for that argument.  This registry currently holds information for
plugins involved in over 1% of builds over the last 28 days where
there is a compatibility issue over the range of AGP/Gradle versions
supported, other than io.fabric / com.google.firebase.crashlytics,
which is handled differently.  (In practice this means the Kotlin and
AndroidX SafeArgs plugins)

Bug: 170806871
Test: added, extended
Change-Id: I3c28e191309f3ea6587658c5aaf862dd039ca677

GitOrigin-RevId: aa0ea64fc01eb4de1b06a06c687bc8cde0ccb687