[AUA] preserve more of the Gradle distribution url
When upgrading the Gradle version in gradle-wrapper.properties,
attempt to preserve the existing structure of the url.
At present, most of the time this delegates to the static method,
which itself hard-codes the Gradle web server and forces an explicit
choice of binary vs combined distribution.  However, for distribution
urls which don't look like the standard Gradle web server download,
attempt to update just the version, in the hope that the new version
will be found in the same place.

(I believe that this is necessary for integration tests to work in a
sandboxed environment.)

Bug: n/a
Test: added
Change-Id: I37dd9308ff441b3ec5c3c98f4bf654bf26a18190
(cherry picked from commit d5f7dd87e71722f595f38e685b558dd3179ce1e2)