enhance GradleWrapper.getUpdatedDistributionUrl
When computing an updated distribution URL for a new gradle version,
preserve as much of the existing distributionUrl property, if any, as
possible.  In particular, if the existing URL is at the canonical
services.gradle.org host, preserve the existing -bin or -all suffix.
Only use the second argument where we cannot deduce the existing
choice.

Bug: 170709509
Test: added, extended
Change-Id: Id80b26726f120360d9c64fe741b03de22dbf6e2e

GitOrigin-RevId: 5af4fb185b1849632d2333760c79f510aad126de