[AUA] upgrade Gradle to minimum compatible version
Rather than upgrade to the minimum compatible version for the version
of AGP corresponding to the running version of Studio, upgrade to the
minimum compatible version for the version of AGP being upgraded to.
(In interactive use these are currently the same, but in tests and for
future functionality they need to be decoupled).

Build a table and a helper accessor to look up the minimum compatible
version for AGP versions that we support, based on
https://developer.android.com/studio/releases/gradle-plugin#updating-gradle
(handling development versions and point releases of studio as the
same as the corresponding .0 release).

Having this functionality means that some component refactoring
processor constructors no longer need the Gradle GradleVersion to be
passed, as they can look up the relevant version.  Adjust the
constructors, and in the process of using the look up function address
existing TODO items.

This also allows a simplification in AndroidPluginVersionUpdater.

Bug: 170803590
Test: existing tests extended and adjusted; unit test added
Change-Id: I245d540c25502cda8d96156b45ca63e298d9bca5

GitOrigin-RevId: b8b0fed289cc17270a0b1c0cfa254725970c48e8