Add option to not show build output on error
The new "build on save" feature on the Compose preview will trigger a
build when the user saves. In these cases, we do not want the output
panel to automatically pop up since the user did not explicely trigger
the build.
This allows for thos build requests to not show the panel automatically.

Bug: 169940598
Test: Manually tested
Change-Id: I0bf560419eb77a8931b49530047f170690f98ee1

GitOrigin-RevId: 2cb31e9ea0be266439e4341ee9d22670771c0bb4