Update config list on dialog close
Config list is updated on PROFILING_CONFIGURATION aspect fired but
so is the dropdown list. Since the order of aspect firing is
undeterministic, we can't rely on one happening before the other.
Thus, we manually update the config list on dialog close to make
sure the dropdown list doesn't get a stale list.

Fixes: 170801047
Test: added
Change-Id: I8d0f838d59e8c326782c72d772cb74c743eec79f