Fix lock wait timeout in ADB initialization
Due to timeout, synchronized, and lock interactions, ADB can get
into a state where if adb.exe does not answer to a verification
request, it would lead to a cascade of lock waits, eventually
stalling the UI thread. Although this would eventually resolve
due to (long) timeouts, it is still unacceptable for the UI thread
to freeze, thereby freezing Studio.

This fix removes the initialization lock as well as all uses of
synchronize. Behind the scenes, the fix is to use a
sequential executor to serialize order all interactions with
AndroidDebugBridge. This allows the removal of synchronized
keywork since all tasks are pushed to a single thread.
Furthermore, the manual lock for interaction with the underlying
AndroidDebugBridge was removed, since init/shutdown are now
always serialized with respect to before-after semantics.

The combination of the lock removal as well as all "sychronized"
erradication is what removes the potential lock wait timeout.

Fixes: 166546817
Test: AdbServiceTest, TerminateAdbIfNotUsedTest, manual testing
Change-Id: I725fa216a9ff6d802fea74337aaed1ea54ac6a6b

GitOrigin-RevId: c75621323f71ee642c64e6eb7721374484fd8141