Fix long freeze in ShowLicensesUsedAction
The licenses use HTML to render and that is very slow to measure. This
CL removes the use of the JTextPane with a label that is much faster to
render.
This reduces the time to display from 20-30s to 2-3s.

Test: Manually verified
Fixes: 170462711
Change-Id: I5a7a52b3f2c935aa8e90c4d3ab367ec1d14c7d12
(cherry picked from commit b9ae39960504abc5f6072f0056fc1ede8849bf4a)