Guard against special identifier in PropKeysLabelInspection
When visiting KtCallExpressions while adding the highlight for lint warnings
corresponding to the absence of the `label` parameter in Compose animation
PropKeys, we were trying to access FqName#shortName() directly. However,
some KtCallExpressions are special (e.g. "<root>") and do not have a short
name. Calling shortName() for those will cause an Exception to be thrown, so
we first need to check if the expression's FqName is special before getting
the short name.

Fixes: 171703620
Test: AnimationInspectionsTest, SantaTrackerKotlinBenchmark
Change-Id: I9ea815dcd57da95623db0ca70931c8ea3f45b608