Detect Gradle version in UnsupportedGradleVersion
... issue checker on messages that have more than a single line.

When we get error messages related to requiring a specific Gradle
version, we look for a pattern and if found this message is slightly
modified by adding new lines.

When looking for the Gradle version to apply, we look again in the
modified message but now the pattern does not match, causing the current
version to be used instead. Since the fix message does not specify what
version will be used, the message implies that it will apply the
recommended version but instead it applies the latest available version
(specified in SdkConstants.GRADLE_LATEST_VERSION).

This change modifies the pattern to consume all lines and not only the
first one and also changes the quick fix message to explicitly say what
version of Gradle will be applied.

Bug: 171583343
Test: UnsupportedGradleVersionIssueCheckerTest
Change-Id: Iedb279fbb76b220be8ba1eb19f1510407ab922c5
(cherry picked from commit 27cfc62d37b03f9881cb24094877e50b73cafcff)