Change event processing to use a sequential coroutine dispatcher
Currently, we're using a worker thread pool executor, which polls gRPC
messages in unspecified order. In practice, this is probably unlikely to
cause an issue, but if an inspector sends multiple events in a row, they
probably expect them to be handled in that order, so it's possible we're
sitting on a potential race condition landmine.

This will matter even more in a near future CL, as we plan to add
sending data via sequential chunks, so processing them out of order
would be very bad.

Bug: N/A
Test: N/A
Change-Id: I091fb1d3f41fc8a82fd33362111592fdc63a8f35

GitOrigin-RevId: 80182f0bcdcffebd3542d90d9eefcca2eefd7635