Some changes around tree updates
Previously we were sometimes calling the skia parser service on the UI
thread. This was because LayoutInspector.loadComponentTree was
needlessly calling layoutInspectorModel.update on the UI thread.
However, the fact that everything was running on the UI thread meant
that we didn't have to worry about synchronization. Now we do, so adding
a mechanism for that.

In DeviceViewContentPanel, we were previously listening for changes on
the model and then refreshing the view. At the time the model is
refreshed we won't necessarily have updated images available yet, and
so refreshing the view at that point isn't useful. Now we wait until
images are loaded to refresh.

Fixes: 170236756
Test: This should be covered by @Slow annotation checking in UI tests.
Test: Unfortunately UI tests aren't being run right now, so we'll have
Test: to check this once they are again (see bug 171430784)
Change-Id: I2c413592f0f1e9087ed45b97cff5689776a4c4d0

GitOrigin-RevId: 8a749ccfd3dc482a6d368c9c86800abda2f2efed