Fix legacy inspector
There were two problems that caused the main view in the pre-api-29
inspector to break:
- In I0fcf7653b, LegacyTreeLoader was incorrectly creating
AndroidWindows with no ability to update drawChildren. This meant that
after an update drawChildren were not repopulated and the view showed as
blank.
- ddmClient.captureView seems broken sometimes. For example, in
Sunflower it seems to work fine on the first page ("Your Garden is
Empty"), but on the "Add Plant" page it returns a broken image, after
which ddm/adb seem to be completely broken. It's not clear what caused
this, but it's now tracked in the context of Layout Inspector by
171901393.

Test: updated and added tests
Bug: none
Change-Id: Ia40d7b28e29bb1cc52f1eae12c0bbcfe79c24536

GitOrigin-RevId: 1c1cba0492eaa2aaae11ce459feadebbbcd7b54d