[AUA] upgrade to at least GRADLE_MINIMUM_VERSION
Although there might be a theoretical compatibility between old AGP
and old Gradle versions, if Studio does not support that old Gradle
version (even if it supports that old AGP version) we should not
upgrade to it.  Compute an in-practice compatible Gradle version for
the AGP upgrade, and use that version for other Gradle plugin
compatibility checks.

Bug: 171583343
Test: existing tests adjusted
Change-Id: I0e3c67a247e7f762d5e811c386c954d7caabf165