[AUA] show a progress bar when parsing
We currently ensure parsing of the build model has been done on a
background thread before showing dialogs, thus preventing significant
UI freezes.  However, for large projects this manifests to the user as
a disconcerting gap in time (around 30 seconds for a 1k-module project
in my experiments) between selecting "Begin Upgrade" and seeing
feedback from the user interface.

Define a new interface function for getting the ProjectBuildModels
with a callback argument, and use it in the Upgrade Assistant to
display the current parsing progress.

This currently leaves a smaller pause before showing the Upgrade
Assistant dialog, where the individual component refactoring
progressors are checked for no-op on the project at hand.  This pause
was around 1 second for a 1k-module project, and could be rectified in
a similar way but with more invasive code changes.

Bug: 172195362
Test: added, manual
Change-Id: Id1b8ebd5ef7dc926de606927ee756cd064e725f4
(cherry picked from commit 8a7ca825d7b05700819172e5d26d8e2d9c252c45)