Fix for PropertyComboBox when popup is not visible
When opening the popup menu, sometimes the 'updatePopup' callback is
invoked before the popup is actually visible, this leads to an Exception
from JPopupMenu#getLocationOnScreen. This likely happens when the model
doesn't have much work to do.

Also fixes a case when the property's initial value is empty, triggering
a model update wirtes the value from EnumValue.LOADING

Bug: n/a
Test: manual, with an example that consistently reproduces the issue
Change-Id: If2598ee660fe61fd4bd36972627d6b9414b81f17

GitOrigin-RevId: e5d90fc26e0d054832e83c32a9759884a4b96c3b