Use FileBasedIndex.getFileData instead of getAllKeys/processValues
The main difference between getFileData and getAllKeys is that with getFileData FileBasedIndexImpl.ensureUpToDate is called with the restrictedFile parameter pointing to AndroidManifest.xml, while with getAllKeys the restrictedFile parameter is null. This means that getFileData forces indexing of only the manifest file while getAllKeys forced indexing of all out of date files.

Test: existing
Bug: 166625311
Bug: 166334891
Change-Id: I41b118bb9894c3dc2fe0ca0ac0c430129f41b70b