Fix duplicated device columns in TestMatrix with apply-code-changes
The apply code changes action instantiates AndroidTestApplicationLaunchTask but skips running
it in the apply-code-changes LaunchContext. Because
onTestSuiteScheduled(device) callback is invoked
from the task's init block, the TestMatrix view
ends up with having duplicated test result columns
for the same device. This patch fixes the issue
by checking the device ID before adding a new
device to the table.

Fixes: 172088812
Test: AndroidTestSuiteViewTest
Change-Id: I1f7a425422044540711d482794b47a76ef908a3a

GitOrigin-RevId: e581a7d21b4e34f84562f993e258388a5bad6730