Make artifact resolver return structured result
This change introduces ArtifactResolverRequest and
ArtifactResolverResult which serve as the new input and output of the
resolver. They act as nice collection objects and also give us
flexibility in the future to add more fields. Most importantly, requests
are bundled in the results which make it very easy to match them.

Test: covered by existing tests.
Bug: n/a
Change-Id: I55d0033b42f6d14977fa54f5fea82744c6b3b162

GitOrigin-RevId: 5c4af71993e174edca4fade09a34766e3e0500c6