[BA] Show suggestion to profile your build
We want to notify users that the default GC was changed in java 9
and as we replace provided java from 8 to 11 now it can impact
the user. Based on our benchmarks default gc has about 10% negative
impact on build speed.
Check for the java vm version and arguments used in gradle daemon
in of build. If java is >=9 and no explicit GC is configured
show the message in Build Analyzer that suggests user to profile
their build to determine the best settings (link to help
documentation provided).

Bug: 163462113
Test: Added
Change-Id: I76a098dd4764932af16356370e22ced877832174

GitOrigin-RevId: e8c7a0094e76f66d69f69a64e86bf697b91c0084