[BA] Allow to hide suggestion to profile your build
Add 'Don't show this again' action link below the notification
warning added in previous cl.
Ask for a confirmation in a message dialog as this action
is not easily undone.
Record suppression information to BuildAttributionWarningsFilter
component that obtains a persistent state. Check that state
when calculating if we need to show the message, don't show
if suppressed.
Hide the message in current report on dialog confirm.

Bug: 163462113
Test: Added
Change-Id: I2a5117b100e19a2f5b15f620ea93773a737198e2

GitOrigin-RevId: 9449bd380317dc98fbe188d06ba1c262beb853b2