LL: Don't track projects / listeners as part of LL Deployment Monitor
Problem: The current implementation attempts to keep a mapping of which project has
LLService activated. This is sometimes out of sync because LLService deactivate itself
sometimes.

Solution: The cleaner / more robust solution is just not to do any of the mapping.
We deactivate when a build starts and activate it when the build succeed.

Fixes: 173434393
Test: None
Change-Id: I85f739b748f90690cf82574572a4aae1f3195a38

GitOrigin-RevId: 32d8fd1d062f53157f7eea9c3dfee4bc7b864581